Remove global variables, made addComment function, rename controller file
